% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sp.R
\name{as_sp}
\alias{as_sp}
\title{Convert a geoVersion object to a Spatial*DataFrame}
\usage{
as_sp(x)
}
\arguments{
\item{x}{the geoVersion object}
}
\description{
Convert a geoVersion object to a Spatial*DataFrame
}

