% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\docType{methods}
\name{combine}
\alias{combine}
\alias{combine,geoVersion-method}
\title{Combine objects}
\usage{
combine(...)

\S4method{combine}{geoVersion}(...)
}
\arguments{
\item{...}{the objects}
}
\description{
Combine objects
}

