% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined.R
\docType{methods}
\name{combined}
\alias{combined}
\alias{combined,geoVersion-method}
\title{Combine objects}
\usage{
combined(...)

\S4method{combined}{geoVersion}(...)
}
\arguments{
\item{...}{the objects}
}
\description{
Combine objects
}

