% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_polygon.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert,Polygon-method}
\alias{convert,Polygons-method}
\alias{convert,SpatialPolygonsDataFrame-method}
\alias{convert,list-method}
\title{Convert a spatial object}
\usage{
convert(object, id)

\S4method{convert}{Polygon}(object, id)

\S4method{convert}{Polygons}(object, id)

\S4method{convert}{SpatialPolygonsDataFrame}(object, id)

\S4method{convert}{list}(object, id)
}
\arguments{
\item{object}{the object to convert}

\item{id}{the name of the variable that holds the stable ID}
}
\description{
Convert a spatial object
}

