% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_db.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve a spatial object from the database}
\usage{
retrieve(name, connection)
}
\arguments{
\item{name}{the name of the layer}

\item{connection}{a DBIConnection}
}
\description{
Retrieve a spatial object from the database
}

