% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation.R
\name{rotation}
\alias{rotation}
\title{calculate the rotation models for a geoVersion object}
\usage{
rotation(object)
}
\arguments{
\item{object}{a geoVersion object}
}
\description{
calculate the rotation models for a geoVersion object
}

