% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform}
\alias{transform}
\title{apply a transformation to a geoVersion object}
\usage{
transform(object, models)
}
\arguments{
\item{object}{a geoVersion object}

\item{models}{a set of transformation models}
}
\description{
apply a transformation to a geoVersion object
}

