% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_crs.R
\name{valid_crs}
\alias{valid_crs}
\title{test if characters represent valid CRS}
\usage{
valid_crs(x)
}
\arguments{
\item{x}{the character vector to test}
}
\description{
test if characters represent valid CRS
}

