% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_polygon.R
\docType{methods}
\name{convert}
\alias{convert}
\alias{convert,Polygon-method}
\title{Convert a spatial object}
\usage{
convert(object)

\S4method{convert}{Polygon}(object)
}
\arguments{
\item{object}{the object to convert}
}
\description{
Convert a spatial object
}

