% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_db.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve a spatial object from the database}
\usage{
retrieve(name, connection, timestamp = NA_real_)
}
\arguments{
\item{name}{the name of the layer}

\item{connection}{a DBIConnection}

\item{timestamp}{The point in time at which the object should be return. Default to NA which implies the latest status.}
}
\description{
Retrieve a spatial object from the database
}

