% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_db.R
\name{store}
\alias{store}
\title{store a geoVersion object}
\usage{
store(x, connection)
}
\arguments{
\item{x}{a geoVersion object}

\item{connection}{a DBIConnection}
}
\description{
store a geoVersion object
}

